# Initial phase of optimizations