- Use slice instead of array for SQL keywords. Go will copy the array every time copying the function - an alternative to slice would be pointer to array, but that is effectively what a slice is while being more idiomatic Go. (HUGE WIN)
- Remove function pointer lookup for calling lookupWord since there is only one implementation anyways
- Switch to stdlib binary search (at the same time prevented uppercasing every single iteration of the search loop)
- Use lookup table for ascii bytes instead of iterating

Still many more easy wins to be found I think, but after finding the HUGE WIN, want to get this in as a start, 30x improvement. More importantly, now CPU profiles actually make sense for finding more problems :D

After
```
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/libinjection-go
BenchmarkSQLiDriver
BenchmarkSQLiDriver/sqli
BenchmarkSQLiDriver/sqli-10         	   24558	     49123 ns/op
BenchmarkSQLiDriver/folding
BenchmarkSQLiDriver/folding-10      	   10000	    103913 ns/op
BenchmarkSQLiDriver/tokens
BenchmarkSQLiDriver/tokens-10       	    6902	    170414 ns/op
PASS
```

Before
```
goos: darwin
goarch: arm64
pkg: github.com/corazawaf/libinjection-go
BenchmarkSQLiDriver
BenchmarkSQLiDriver/sqli
BenchmarkSQLiDriver/sqli-10         	     620	   1882274 ns/op
BenchmarkSQLiDriver/folding
BenchmarkSQLiDriver/folding-10      	     484	   2464673 ns/op
BenchmarkSQLiDriver/tokens
BenchmarkSQLiDriver/tokens-10       	     368	   3258661 ns/op
```