@anuraaga 2022-09-07
@jptosso @jcchavezs Can you have a look at this? The performance fix here is P0 I think