@jptosso 2022-09-08
LGTM, I think the linter is failing because we should switch from "git" to "https"